Extract link checker ignores to a dedicated file
